+++ clients/admin-ui/cypress/e2e/consent-configuration.cy.ts
@@ -2,6 +2,7 @@ import {
   stubPlus,
   stubSystemCrud,
   stubTaxonomyEntities,
+  stubVendorList,
 } from "cypress/support/stubs";
 
 import { CONFIGURE_CONSENT_ROUTE } from "~/features/common/nav/v2/routes";
@@ -70,6 +71,612 @@ describe("Consent configuration", () => {
 
       cy.getByTestId("add-cookie-btn");
       cy.getByTestId("add-vendor-btn");
+    });
+  });
+
+  describe("adding a vendor", () => {
+    beforeEach(() => {
+      stubSystemCrud();
+      stubTaxonomyEntities();
+      stubVendorList();
+      cy.intercept("GET", "/api/v1/system", {
+        fixture: "systems/systems.json",
+      }).as("getSystems");
+    });
+
+    describe("without the dictionary", () => {
+      beforeEach(() => {
+        stubPlus(true, {
+          core_fides_version: "1.9.6",
+          fidesplus_server: "healthy",
+          fidesplus_version: "1.9.6",
+          system_scanner: {
+            enabled: false,
+            cluster_health: null,
+            cluster_error: null,
+          },
+          dictionary: {
+            enabled: false,
+            service_health: null,
+            service_error: null,
+          },
+          fides_cloud: {
+            enabled: false,
+          },
+        });
+        cy.visit(CONFIGURE_CONSENT_ROUTE);
+      });
+
+      it("can add a vendor without the dictionary", () => {
+        cy.getByTestId("add-vendor-btn").click();
+        cy.getByTestId("input-name").type("test vendor");
+        cy.selectOption("input-privacy_declarations.0.data_use", "analytics");
+        cy.getByTestId("input-privacy_declarations.0.cookieNames")
+          .find(".custom-creatable-select__input-container")
+          .type("test{enter}cookie{enter}");
+        cy.getByTestId("save-btn").click();
+        cy.wait("@postSystem").then((interception) => {
+          const { body } = interception.request;
+          expect(body).to.eql({
+            name: "test vendor",
+            fides_key: "test_vendor",
+            system_type: "",
+            privacy_declarations: [
+              {
+                name: "",
+                data_use: "analytics",
+                data_categories: ["user"],
+                cookies: [
+                  {
+                    name: "test",
+                    path: "/",
+                  },
+                  {
+                    name: "cookie",
+                    path: "/",
+                  },
+                ],
+              },
+            ],
+          });
+        });
+      });
+
+      it("can manually add more data uses", () => {
+        cy.getByTestId("add-vendor-btn").click();
+        cy.getByTestId("add-data-use-btn").should("be.disabled");
+        cy.getByTestId("input-name").type("test vendor");
+        cy.selectOption("input-privacy_declarations.0.data_use", "analytics");
+        cy.getByTestId("input-privacy_declarations.0.cookieNames")
+          .find(".custom-creatable-select__input-container")
+          .type("one{enter}");
+
+        // Add another use
+        cy.getByTestId("add-data-use-btn").click();
+        cy.selectOption(
+          "input-privacy_declarations.1.data_use",
+          "personalize.system"
+        );
+        cy.getByTestId("input-privacy_declarations.1.cookieNames")
+          .find(".custom-creatable-select__input-container")
+          .type("two{enter}");
+        cy.getByTestId("save-btn").click();
+        cy.wait("@postSystem").then((interception) => {
+          const { body } = interception.request;
+          expect(body.privacy_declarations).to.eql([
+            {
+              name: "",
+              data_use: "analytics",
+              data_categories: ["user"],
+              cookies: [
+                {
+                  name: "one",
+                  path: "/",
+                },
+              ],
+            },
+            {
+              name: "",
+              data_use: "personalize.system",
+              data_categories: ["user"],
+              cookies: [
+                {
+                  name: "two",
+                  path: "/",
+                },
+              ],
+            },
+          ]);
+        });
+      });
+    });
+
+    describe("with the dictionary", () => {
+      beforeEach(() => {
+        stubPlus(true);
+        cy.visit(CONFIGURE_CONSENT_ROUTE);
+      });
+
+      it("can fill in dictionary suggestions", () => {
+        cy.getByTestId("add-vendor-btn").click();
+        cy.getByTestId("input-vendor_id")
+          .click()
+          .find(`.custom-creatable-select__menu-list`)
+          .contains("Aniview LTD")
+          .click();
+        cy.getByTestId("sparkle-btn").click();
+        cy.wait("@getDictionaryDeclarations");
+        cy.getSelectValueContainer(
+          "input-privacy_declarations.0.data_use"
+        ).contains("marketing.advertising.profiling");
+        ["av_*", "aniC", "2_C_*"].forEach((cookieName) => {
+          cy.getByTestId("input-privacy_declarations.0.cookieNames").contains(
+            cookieName
+          );
+        });
+
+        // Also check one that shouldn't have any cookies
+        cy.getSelectValueContainer(
+          "input-privacy_declarations.3.data_use"
+        ).contains("analytics.reporting.campaign_insights");
+        cy.getByTestId("input-privacy_declarations.3.cookieNames").contains(
+          "Select..."
+        );
+        // There should be 13 declarations (but start from 0, so 12)
+        cy.getByTestId("input-privacy_declarations.12.cookieNames");
+        cy.getByTestId("save-btn").click();
+        cy.wait("@postSystem").then((interception) => {
+          const { body } = interception.request;
+          expect(body.name).to.eql("Aniview LTD");
+          expect(body.privacy_declarations).to.eql([
+            {
+              name: "",
+              data_use: "marketing.advertising.profiling",
+              data_categories: [
+                "user.device.ip_address",
+                "user.device",
+                "user.sensor",
+                "user.user_sensor",
+                "user.telemetry",
+                "user.device.cookie_id",
+                "user.device.device_id",
+                "user.device.cookie",
+                "user.behavior.purchase_history",
+                "user.behavior",
+                "user.behavior.browsing_history",
+                "user.behavior.media_consumption",
+                "user.behavior.search_history",
+                "user.social",
+                "user.location.imprecise",
+              ],
+              cookies: [
+                {
+                  name: "av_*",
+                  domain: "**",
+                  path: "/",
+                },
+                {
+                  name: "aniC",
+                  domain: "*.aniview.com*.aniview.com",
+                  path: "/",
+                },
+                {
+                  name: "2_C_*",
+                  domain: "*.aniview.com*.aniview.com",
+                  path: "/",
+                },
+              ],
+            },
+            {
+              name: "",
+              data_use: "functional.storage",
+              data_categories: [
+                "user.device.ip_address",
+                "user.device",
+                "user.sensor",
+                "user.user_sensor",
+                "user.telemetry",
+                "user.device.cookie_id",
+                "user.device.device_id",
+                "user.device.cookie",
+                "user.behavior.purchase_history",
+                "user.behavior",
+                "user.behavior.browsing_history",
+                "user.behavior.media_consumption",
+                "user.behavior.search_history",
+                "user.social",
+                "user.location.imprecise",
+              ],
+              cookies: [
+                {
+                  name: "av_*",
+                  domain: "**",
+                  path: "/",
+                },
+                {
+                  name: "aniC",
+                  domain: "*.aniview.com*.aniview.com",
+                  path: "/",
+                },
+                {
+                  name: "2_C_*",
+                  domain: "*.aniview.com*.aniview.com",
+                  path: "/",
+                },
+              ],
+            },
+            {
+              name: "",
+              data_use: "analytics.reporting.ad_performance",
+              data_categories: [
+                "user.device.ip_address",
+                "user.device",
+                "user.sensor",
+                "user.user_sensor",
+                "user.telemetry",
+                "user.device.cookie_id",
+                "user.device.device_id",
+                "user.device.cookie",
+                "user.behavior.purchase_history",
+                "user.behavior",
+                "user.behavior.browsing_history",
+                "user.behavior.media_consumption",
+                "user.behavior.search_history",
+                "user.social",
+                "user.location.imprecise",
+              ],
+              cookies: [
+                {
+                  name: "av_*",
+                  domain: "**",
+                  path: "/",
+                },
+                {
+                  name: "aniC",
+                  domain: "*.aniview.com*.aniview.com",
+                  path: "/",
+                },
+                {
+                  name: "2_C_*",
+                  domain: "*.aniview.com*.aniview.com",
+                  path: "/",
+                },
+              ],
+            },
+            {
+              name: "",
+              data_use: "analytics.reporting.campaign_insights",
+              data_categories: [
+                "user.device.ip_address",
+                "user.device",
+                "user.sensor",
+                "user.user_sensor",
+                "user.telemetry",
+                "user.device.cookie_id",
+                "user.device.device_id",
+                "user.device.cookie",
+                "user.behavior.purchase_history",
+                "user.behavior",
+                "user.behavior.browsing_history",
+                "user.behavior.media_consumption",
+                "user.behavior.search_history",
+                "user.social",
+                "user.location.imprecise",
+              ],
+              cookies: [],
+            },
+            {
+              name: "",
+              data_use: "marketing.advertising.first_party.targeted",
+              data_categories: [
+                "user.device.ip_address",
+                "user.device",
+                "user.sensor",
+                "user.user_sensor",
+                "user.telemetry",
+                "user.device.cookie_id",
+                "user.device.device_id",
+                "user.device.cookie",
+                "user.behavior.purchase_history",
+                "user.behavior",
+                "user.behavior.browsing_history",
+                "user.behavior.media_consumption",
+                "user.behavior.search_history",
+                "user.social",
+                "user.location.imprecise",
+              ],
+              cookies: [
+                {
+                  name: "av_*",
+                  domain: "**",
+                  path: "/",
+                },
+                {
+                  name: "aniC",
+                  domain: "*.aniview.com*.aniview.com",
+                  path: "/",
+                },
+                {
+                  name: "2_C_*",
+                  domain: "*.aniview.com*.aniview.com",
+                  path: "/",
+                },
+              ],
+            },
+            {
+              name: "",
+              data_use: "functional.service.improve",
+              data_categories: [
+                "user.device.ip_address",
+                "user.device",
+                "user.sensor",
+                "user.user_sensor",
+                "user.telemetry",
+                "user.device.cookie_id",
+                "user.device.device_id",
+                "user.device.cookie",
+                "user.behavior.purchase_history",
+                "user.behavior",
+                "user.behavior.browsing_history",
+                "user.behavior.media_consumption",
+                "user.behavior.search_history",
+                "user.social",
+                "user.location.imprecise",
+              ],
+              cookies: [],
+            },
+            {
+              name: "",
+              data_use: "analytics.reporting.content_performance",
+              data_categories: [
+                "user.device.ip_address",
+                "user.device",
+                "user.sensor",
+                "user.user_sensor",
+                "user.telemetry",
+                "user.device.cookie_id",
+                "user.device.device_id",
+                "user.device.cookie",
+                "user.behavior.purchase_history",
+                "user.behavior",
+                "user.behavior.browsing_history",
+                "user.behavior.media_consumption",
+                "user.behavior.search_history",
+                "user.social",
+                "user.location.imprecise",
+              ],
+              cookies: [],
+            },
+            {
+              name: "",
+              data_use: "essential.fraud_detection",
+              data_categories: [
+                "user.device.ip_address",
+                "user.device",
+                "user.sensor",
+                "user.user_sensor",
+                "user.telemetry",
+                "user.device.cookie_id",
+                "user.device.device_id",
+                "user.device.cookie",
+                "user.behavior.purchase_history",
+                "user.behavior",
+                "user.behavior.browsing_history",
+                "user.behavior.media_consumption",
+                "user.behavior.search_history",
+                "user.social",
+                "user.location.imprecise",
+              ],
+              cookies: [],
+            },
+            {
+              name: "",
+              data_use: "essential.service.security",
+              data_categories: [
+                "user.device.ip_address",
+                "user.device",
+                "user.sensor",
+                "user.user_sensor",
+                "user.telemetry",
+                "user.device.cookie_id",
+                "user.device.device_id",
+                "user.device.cookie",
+                "user.behavior.purchase_history",
+                "user.behavior",
+                "user.behavior.browsing_history",
+                "user.behavior.media_consumption",
+                "user.behavior.search_history",
+                "user.social",
+                "user.location.imprecise",
+              ],
+              cookies: [],
+            },
+            {
+              name: "",
+              data_use: "marketing.advertising.negative_targeting",
+              data_categories: [
+                "user.device.ip_address",
+                "user.device",
+                "user.sensor",
+                "user.user_sensor",
+                "user.telemetry",
+                "user.device.cookie_id",
+                "user.device.device_id",
+                "user.device.cookie",
+                "user.behavior.purchase_history",
+                "user.behavior",
+                "user.behavior.browsing_history",
+                "user.behavior.media_consumption",
+                "user.behavior.search_history",
+                "user.social",
+                "user.location.imprecise",
+              ],
+              cookies: [
+                {
+                  name: "av_*",
+                  domain: "**",
+                  path: "/",
+                },
+                {
+                  name: "aniC",
+                  domain: "*.aniview.com*.aniview.com",
+                  path: "/",
+                },
+                {
+                  name: "2_C_*",
+                  domain: "*.aniview.com*.aniview.com",
+                  path: "/",
+                },
+              ],
+            },
+            {
+              name: "",
+              data_use: "marketing.advertising.first_party.contextual",
+              data_categories: [
+                "user.device.ip_address",
+                "user.device",
+                "user.sensor",
+                "user.user_sensor",
+                "user.telemetry",
+                "user.device.cookie_id",
+                "user.device.device_id",
+                "user.device.cookie",
+                "user.behavior.purchase_history",
+                "user.behavior",
+                "user.behavior.browsing_history",
+                "user.behavior.media_consumption",
+                "user.behavior.search_history",
+                "user.social",
+                "user.location.imprecise",
+              ],
+              cookies: [
+                {
+                  name: "av_*",
+                  domain: "**",
+                  path: "/",
+                },
+                {
+                  name: "aniC",
+                  domain: "*.aniview.com*.aniview.com",
+                  path: "/",
+                },
+                {
+                  name: "2_C_*",
+                  domain: "*.aniview.com*.aniview.com",
+                  path: "/",
+                },
+              ],
+            },
+            {
+              name: "",
+              data_use: "marketing.advertising.frequency_capping",
+              data_categories: [
+                "user.device.ip_address",
+                "user.device",
+                "user.sensor",
+                "user.user_sensor",
+                "user.telemetry",
+                "user.device.cookie_id",
+                "user.device.device_id",
+                "user.device.cookie",
+                "user.behavior.purchase_history",
+                "user.behavior",
+                "user.behavior.browsing_history",
+                "user.behavior.media_consumption",
+                "user.behavior.search_history",
+                "user.social",
+                "user.location.imprecise",
+              ],
+              cookies: [
+                {
+                  name: "av_*",
+                  domain: "**",
+                  path: "/",
+                },
+                {
+                  name: "aniC",
+                  domain: "*.aniview.com*.aniview.com",
+                  path: "/",
+                },
+                {
+                  name: "2_C_*",
+                  domain: "*.aniview.com*.aniview.com",
+                  path: "/",
+                },
+              ],
+            },
+            {
+              name: "",
+              data_use: "marketing.advertising.third_party.targeted",
+              data_categories: [
+                "user.device.ip_address",
+                "user.device",
+                "user.sensor",
+                "user.user_sensor",
+                "user.telemetry",
+                "user.device.cookie_id",
+                "user.device.device_id",
+                "user.device.cookie",
+                "user.behavior.purchase_history",
+                "user.behavior",
+                "user.behavior.browsing_history",
+                "user.behavior.media_consumption",
+                "user.behavior.search_history",
+                "user.social",
+                "user.location.imprecise",
+              ],
+              cookies: [
+                {
+                  name: "av_*",
+                  domain: "**",
+                  path: "/",
+                },
+                {
+                  name: "aniC",
+                  domain: "*.aniview.com*.aniview.com",
+                  path: "/",
+                },
+                {
+                  name: "2_C_*",
+                  domain: "*.aniview.com*.aniview.com",
+                  path: "/",
+                },
+              ],
+            },
+          ]);
+        });
+      });
+
+      it("can create a vendor that is not in the dictionary", () => {
+        cy.getByTestId("add-vendor-btn").click();
+        cy.getByTestId("input-vendor_id").type("custom vendor{enter}");
+        cy.selectOption("input-privacy_declarations.0.data_use", "analytics");
+        cy.getByTestId("input-privacy_declarations.0.cookieNames")
+          .find(".custom-creatable-select__input-container")
+          .type("test{enter}");
+        cy.getByTestId("save-btn").click();
+        cy.wait("@postSystem").then((interception) => {
+          const { body } = interception.request;
+          expect(body).to.eql({
+            name: "custom vendor",
+            fides_key: "custom_vendor",
+            system_type: "",
+            privacy_declarations: [
+              {
+                name: "",
+                data_use: "analytics",
+                data_categories: ["user"],
+                cookies: [
+                  {
+                    name: "test",
+                    path: "/",
+                  },
+                ],
+              },
+            ],
+          });
+        });
+      });
     });
   });
 });